requirements.txt
Flask==0.9
Flask-Login==0.1.3
Flask-Principal==0.3.3
Flask-SQLAlchemy==0.16
Flask-SSLify==0.1.4
Flask-Testing==0.4
Flask-WTF==0.8.2
Jinja2==2.6
SQLAlchemy==0.7.9
WTForms==1.0.2
Werkzeug==0.8.3
blinker==1.2
distribute==0.6.32
lxml==3.1.0
py-bcrypt==0.2
twill==0.9
wsgiref==0.1.2
